+++ src/Adapter/IntendedDegreeProgram/IntendedDegreeProgram.mjs
@@ -2,5 +2,5 @@
 /** @typedef {import("../Subject/Subject.mjs").Subject} Subject */
 
 /**
- * @typedef {{subjects: Subject[], values?: ChosenIntendedDegreeProgram | null}} IntendedDegreeProgram
+ * @typedef {{subjects: Subject[], values: ChosenIntendedDegreeProgram | null}} IntendedDegreeProgram
  */